implements card groups for events
