+++ src/dw_design_system/dwds/components/banner_card.scss
@@ -1,23 +1,11 @@
 .dwds-card.dwds-banner-card {
     h3 {
-        color: var(--color-text-highlight);
         font-size: var(--text-xlarge);
         font-weight: var(--font-weight-normal);
         line-height: var(--font-lineheight-headings);
     }
 
     p {
-        color: var(--color-text-highlight);
         margin: 0;
-    }
-
-    a,
-    a:visited,
-    a:link {
-        color: var(--color-text-highlight);
-
-        p {
-            color: var(--color-text-highlight);
-        }
     }
 }
