+++ src/dw_design_system/dwds/components/card_group.scss
@@ -18,4 +18,9 @@
             }
         }
     }
+
+}
+
+.dwds-card.dwds-card-group.dwds-card-group-grid {
+    // TODO: Add events grid logic here
 }
