+++ src/dw_design_system/dwds/components/cta_card.html
@@ -1,6 +1,6 @@
 {% extends "dwds/elements/extendable/card.html" %}
 
-{% block card_container_classes %}dwds-cta-card dwds-card-blue{% endblock %}
+{% block card_container_classes %}dwds-cta-card card-without-shadow dwds-card-blue{% endblock %}
 
 {% block card_header %}
 {% endblock card_header %}
