+++ src/dw_design_system/dwds/components/cta_card.scss
@@ -1,14 +1,1 @@
-.dwds-card.dwds-cta-card {
-    color: var(--color-text-highlight);
-
-    box-shadow: none;
-
-    a {
-        color: var(--color-text-highlight);
-        text-decoration-color: var(--color-text-highlight);
-
-        p {
-            color: var(--color-text-highlight);
-        }
-    }
-}
+.dwds-card.dwds-cta-card {}
