+++ src/home/templates/home/home_page_new.html
@@ -4,7 +4,6 @@
 {% load wagtailimages_tags %}
 {% load site_alert %}
 {% load cards %}
-{% load event_tags %}
 
 {% block pre_main %}
     {{ block.super }}
@@ -56,16 +55,18 @@
                 </div>
 
                 {% pages_to_cards news_items as news_cards %}
-                {% include "dwds/components/card_group.html" with card_group_title="Latest DBT news" cards=news_cards footer_link="/news-and-views/" footer_link_text="View all news" %}
+                {% include "dwds/components/card_group.html" with card_group_title="Latest DBT news" cards=news_cards footer_link="/news-and-views/" footer_link_text="View all news" show_hr=True %}
                 
+                {% pages_to_cards events as event_cards %}
+                {% include "dwds/components/card_group.html" with grid=True blue_bg=True card_group_title="Events" cards=event_cards %}
+
                 {% if events %}
                     <div class="stack events-container">
                         <h2 class="events-section-title">Events</h2>
                         <div class="grid events-cards-container">
                             {% for event in events %}
-                                {% get_event_date event as event_date %}
-                                {% get_event_time event as event_time %}
-                                {% include "dwds/components/engagement_card.html" with event=event post_title_date=event_date post_title_time=event_time title=event.title thumbnail=event.preview_image url=event.url excerpt=event.excerpt is_highlighted=True date=event.last_published_at %}
+                                
+                                {% include "dwds/components/engagement_card.html" with  title=event.title thumbnail=event.preview_image url=event.url excerpt=event.excerpt is_highlighted=True %}
                             {% endfor %}
                         </div>
                     </div>
